requirements.txt
argparse
blessed>=1.15.0
bs4
cmd2>=0.9.1
colorama
colored>=1.3.93
colorlog>=3.1.4
enlighten>=1.0.7
humanfriendly>=4.12.1
prettytable>=0.7.2
python-libnmap>=0.7.0
regex>=2018.6.9
requests
six>=1.11.0
SQLAlchemy>=1.2.8
SQLAlchemy-Utils>=0.33.3
urllib3